#5 Fix issue with recursive trees
This issue (and PR #6) uncovered an issue I did not consider - a
an AST with a loop breaks the library in many ways, most likely
by just hitting the recursion limit inside Python.

This commit bans recursion to fix the problem. We can later lift the
restriction in select places (i.e. recursive datatype declarations),
and ONLY in those places.